pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lattice"
version = "0.1.0"
description = "Hierarchical multi-agent orchestration with learned routing, adaptive planning, and formal workflow verification."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Lattice Contributors" },
]
keywords = [
    "multi-agent",
    "orchestration",
    "llm",
    "routing",
    "verification",
    "planning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.5,<3",
    "torch>=2.1",
    "z3-solver>=4.12",
    "redis>=5.0",
    "opentelemetry-api>=1.22",
    "opentelemetry-sdk>=1.22",
    "opentelemetry-exporter-otlp>=1.22",
    "litellm>=1.20",
    "openai>=1.10",
    "anthropic>=0.18",
    "numpy>=1.26",
    "tiktoken>=0.5",
    "structlog>=24.1",
    "tenacity>=8.2",
    "networkx>=3.2",
    "aiofiles>=23.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-timeout>=2.2",
    "ruff>=0.2",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]

[project.urls]
Homepage = "https://github.com/lattice-org/lattice"
Documentation = "https://lattice-org.github.io/lattice"
Repository = "https://github.com/lattice-org/lattice"
Issues = "https://github.com/lattice-org/lattice/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/lattice"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
